week05 assignment

+++ Week05/assignment/SSSF-week4-5/src/api/resolvers/catResolver.ts
@@ -58,19 +58,53 @@ export default {
   },
 
   Mutation: {
-    createCat: async (_parent: any, args: Cat, _context: any, _info: any) => {
-      const newCat = new catModel(args);
-      return await newCat.save();
+    createCat: async (_parent: any, args: any, user: UserIdWithToken) => {
+      if (!user.token) {
+        throw new GraphQLError('Not authorized', {
+          extensions: {code: 'NOT_AUTHORIZED'},
+        });
+      }
+
+      args.owner = user.id as unknown as Types.ObjectId;
+
+      const cat = await catModel.create(args);
+      return cat;
     },
 
-    updateCat: async (_parent: any, args: Cat, _context: any, _info: any) => {
+    updateCat: async (_parent: any, args: Cat, user: UserIdWithToken) => {
+      if (!user.token) {
+        throw new GraphQLError('Not authorized', {
+          extensions: {code: 'NOT_AUTHORIZED'},
+        });
+      }
+
+      const found = await catModel.findById(args.id);
+      if (found!.owner.toString() !== user.id.toString()) {
+        throw new GraphQLError('Not authorized', {
+          extensions: {code: 'NOT_AUTHORIZED'},
+        });
+      }
+
       const cat = await catModel.findByIdAndUpdate(args.id, args, {
         new: true,
       });
       return cat;
     },
 
-    deleteCat: async (_parent: any, args: Cat, _context: any, _info: any) => {
+    deleteCat: async (_parent: any, args: Cat, user: UserIdWithToken) => {
+      if (!user.token) {
+        throw new GraphQLError('Not authorized', {
+          extensions: {code: 'NOT_AUTHORIZED'},
+        });
+      }
+
+      const found = await catModel.findById(args.id);
+      if (found!.owner.toString() !== user.id.toString()) {
+        throw new GraphQLError('Not authorized', {
+          extensions: {code: 'NOT_AUTHORIZED'},
+        });
+      }
+
       const cat = await catModel.findByIdAndDelete(args.id);
       return cat;
     },
@@ -78,8 +112,7 @@ export default {
     updateCatAsAdmin: async (
       _parent: any,
       args: Cat,
-      context: {user: UserIdWithToken},
-      _info: any
+      user: UserIdWithToken
     ) => {
       const cat = await catModel.findByIdAndUpdate(args.id, args, {
         new: true,
@@ -90,8 +123,7 @@ export default {
     deleteCatAsAdmin: async (
       _parent: any,
       args: Cat,
-      context: {user: UserIdWithToken},
-      _info: any
+      user: UserIdWithToken
     ) => {
       const cat = await catModel.findByIdAndDelete(args.id);
       return cat;
